+++ pkg/storage/export/export/BUILD.bazel
@@ -3,16 +3,16 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["export.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-controller/watch/export",
+    importpath = "kubevirt.io/kubevirt/pkg/storage/export/export",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/certificates/bootstrap:go_default_library",
         "//pkg/certificates/triple:go_default_library",
         "//pkg/certificates/triple/cert:go_default_library",
         "//pkg/controller:go_default_library",
+        "//pkg/storage/snapshot:go_default_library",
         "//pkg/storage/types:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
-        "//pkg/virt-controller/watch/snapshot:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//pkg/virt-operator/resource/generate/components:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
