+++ pkg/storage/export/virt-exportserver/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["exportserver.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-exportserver",
+    importpath = "kubevirt.io/kubevirt/pkg/storage/export/virt-exportserver",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/service:go_default_library",
