+++ pkg/virt-controller/watch/clone/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-controller/watch/clone",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/storage/snapshot:go_default_library",
         "//pkg/util/status:go_default_library",
-        "//pkg/virt-controller/watch/snapshot:go_default_library",
         "//staging/src/kubevirt.io/api/clone:go_default_library",
         "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
